Renamed dcga to dg3c to match g3c
