+++ clifford/dg3c.py
